Check whether current CWD is a symlink when normalizing paths.
If terraform init is run in a symlinked directory at a different level in the hierarchy than its symlinked target, terraform will fail to resolve the relative path generated based on the target. This commit checks whether the current working directory is a symlink and builds the relative path based on the result.