CARRY: backend: drop un-used remote backeds in terraform
Currently these backends force the installer to vendor in these libraries
forcing the installer to deal with old cloud libraries.

The installer does not use these remote backends anyway, so this should help
a lot to reduce the un-used codepaths and allow installer to use cloud libraries tied to
the providers instead.